fix(eval): cast bf16 actions to fp32 before numpy conversion
control_step_from_obs now ensures action_env and action_model are fp32
before returning. eval_libero_rollout.py also adds defensive .float()
before .cpu().numpy(). This prevents TypeError when model weights are
bf16 (numpy has no bfloat16 support).

+++ libero_hybrid/scripts/eval_libero_rollout.py
@@ -167,7 +167,8 @@ def evaluate_task(
                 )
 
                 action = step_out.action_env[0]  # [A]
-                actions_batch[k] = action.cpu().numpy()
+                # Belt-and-suspenders: ensure fp32 before numpy (bf16 has no numpy equiv)
+                actions_batch[k] = action.float().cpu().numpy()
 
             obs, reward, done, info = env.step(actions_batch)
             total_steps += 1
